Flask router 구현: / -> main.html -> /learnchart
syriness 코드(flask 앱)와 lazychoi 코드(dash 앱) 연결

1. routers.py로  app.py의 라우팅 코드 이동
2. 파라미터 입력 폼에서 submit 버튼 클릭하면 학습 결과 그래프 화면 출력하는 주소(/learnchart)로 이동

해결해야 할 과제
1. 파라미터 입력 폼 예외를 main.html 자체에서 최대한 처리
2. db 접속 최소화
3. rltrader 코드의 CLI 입력를 삭제하고 폼에서 가져온 변수값을 rltrader에 바로 전달
